svg/struct/scripted/svg-checkIntersection-001.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><script>const test = async_test(&apos;checkIntersection() / checkEnclosure() with null ref. element&apos;); window.onload = test.step_func_done(() =&gt; { const root = document.documentElement; assert_throws_js(TypeError, () =&gt; root.checkIntersection(null, root.createSVGRect())); assert_throws_js(TypeError, () =&gt; root.checkEnclosure(null, root.createSVGRect())); });</script><text>PASS</text></svg>
